Update veckoschemar.md
Here is my proposal of change of planning schedule so you can see who is available (my available weeks filled in that I will try to stick to). Christoffer can edit where he is available and then it is easier to decide who should help when as fare as possible. See you tomorrow!